choozi - Recombine code for B1/B2 and use setUI
